+++ cmd/cmd.go
@@ -101,7 +101,7 @@ func initSrv(cfg *config.Config) {
 	for name, item := range cfg.Forms {
 		rls[name] = item.Ratelimit
 	}
-	v := validator.New(cfg.Spam.Hosts, cfg.Spam.Emails, cfg.LogLevel)
+	v := validator.New(cfg.Spam.Hosts, cfg.Spam.Emails, cfg.Spam.Localparts, cfg.LogLevel)
 	pm := mail.New(cfg.Postmark.Token, cfg.Postmark.From, cfg.Postmark.ReplyTo, cfg.LogLevel)
 	fh := sub.NewHandler(cfg.Forms, v, pm, mxb, cfg.LogLevel)
 	srv = web.New(cfg.Port, rls, cfg.LogLevel, fh, v, cfg.Ban.Duration, cfg.Ban.Size)