+++ sub/handler_test.go
@@ -25,7 +25,7 @@ func (s *HandlerSuite) SetupTest() {
 }
 
 func (s *HandlerSuite) TestNew() {
-	handler := NewHandler(nil, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(nil, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 
 	s.IsType(&Handler{}, handler)
 }
@@ -37,7 +37,7 @@ func (s *HandlerSuite) TestGet() {
 			Redirect: "https://example.com",
 		},
 	}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 
 	result, err := handler.GET("test", nil)
 
@@ -47,7 +47,7 @@ func (s *HandlerSuite) TestGet() {
 
 func (s *HandlerSuite) TestGet_NoForm() {
 	forms := map[string]*config.Form{}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 
 	result, err := handler.GET("test", nil)
 
@@ -58,7 +58,7 @@ func (s *HandlerSuite) TestGet_NoForm() {
 
 func (s *HandlerSuite) TestPOST_NoForm() {
 	forms := map[string]*config.Form{}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 
 	result, err := handler.POST("test", nil)
 
@@ -70,7 +70,7 @@ func (s *HandlerSuite) TestPOST_NoForm() {
 func (s *HandlerSuite) TestPOST_NoData() {
 	expected := "<html><head><title>Redirecting...</title><meta http-equiv=\"Refresh\" content=\"0; url='https://example.com'\" /></head><body>Redirecting to <a href='https://example.com'>https://example.com</a>..."
 	forms := map[string]*config.Form{"test": {Redirect: "https://example.com"}}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 	request, rerr := http.NewRequest("POST", "", nil)
 
 	result, err := handler.POST("test", request)
@@ -83,7 +83,7 @@ func (s *HandlerSuite) TestPOST_NoData() {
 func (s *HandlerSuite) TestPOST_SpamEmail() {
 	expected := "<html><head><title>Redirecting...</title><meta http-equiv=\"Refresh\" content=\"0; url='https://example.com'\" /></head><body>Redirecting to <a href='https://example.com'>https://example.com</a>..."
 	forms := map[string]*config.Form{"test": {Redirect: "https://example.com"}}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 	data := url.Values{}
 	data.Add("email", "no")
 	request, rerr := http.NewRequest("POST", "", strings.NewReader(data.Encode()))
@@ -100,7 +100,7 @@ func (s *HandlerSuite) TestPOST_SpamEmail() {
 func (s *HandlerSuite) TestPOST_SpamDomain() {
 	expected := "<html><head><title>Redirecting...</title><meta http-equiv=\"Refresh\" content=\"0; url='https://example.com'\" /></head><body>Redirecting to <a href='https://example.com'>https://example.com</a>..."
 	forms := map[string]*config.Form{"test": {Redirect: "https://example.com"}}
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 	data := url.Values{}
 	data.Add("domain", "no")
 	request, rerr := http.NewRequest("POST", "", strings.NewReader(data.Encode()))
@@ -128,7 +128,7 @@ func (s *HandlerSuite) TestPOST() {
 		},
 	}
 	s.sender.On("Send", roomID, expectedMessage).Once()
-	handler := NewHandler(forms, validator.New([]string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
+	handler := NewHandler(forms, validator.New([]string{}, []string{}, []string{}, "TRACE"), nil, s.sender, "TRACE")
 	data := url.Values{}
 	data.Add("email", "test@example.com")
 	data.Add("field", "value")